Added a few things to BufferPipe.

+++ ManuvrOS/LogPipe.h
@@ -36,8 +36,8 @@ class LogPipe : public BufferPipe {
     ~LogPipe();
 
     /* Override from BufferPipe. */
-    virtual unsigned int toCounterparty(uint8_t* buf, unsigned int len, int8_t mm);
-    virtual unsigned int fromCounterparty(uint8_t* buf, unsigned int len, int8_t mm);
+    virtual int8_t toCounterparty(uint8_t* buf, unsigned int len, int8_t mm);
+    virtual int8_t fromCounterparty(uint8_t* buf, unsigned int len, int8_t mm);
 
     void printDebug(StringBuilder*);
