+++ ManuvrOS/Transports/BufferPipes/ManuvrTLS/ManuvrTLS.h
@@ -60,8 +60,8 @@ class ManuvrTLS : protected BufferPipe {
     ~ManuvrTLS();
 
     /* Override from BufferPipe. */
-    virtual unsigned int toCounterparty(uint8_t* buf, unsigned int len, int8_t mm);
-    virtual unsigned int fromCounterparty(uint8_t* buf, unsigned int len, int8_t mm);
+    virtual int8_t toCounterparty(uint8_t* buf, unsigned int len, int8_t mm);
+    virtual int8_t fromCounterparty(uint8_t* buf, unsigned int len, int8_t mm);
 
 
   protected:
