+++ ManuvrOS/Transports/BufferPipes/XportBridge/XportBridge.h
@@ -53,8 +53,8 @@ class XportBridge : public BufferPipe {
     ~XportBridge();
 
     /* Override from BufferPipe. */
-    virtual unsigned int toCounterparty(uint8_t* buf, unsigned int len, int8_t mm);
-    virtual unsigned int fromCounterparty(uint8_t* buf, unsigned int len, int8_t mm);
+    virtual int8_t toCounterparty(uint8_t* buf, unsigned int len, int8_t mm);
+    virtual int8_t fromCounterparty(uint8_t* buf, unsigned int len, int8_t mm);
 
     virtual void printDebug(StringBuilder*);
